Move Snatchable to its own module
